Add test for new added functions
